+++ src/TurtlePath.Automations/Generation/IAutomationHandlerTypeGeneratorFactory.cs
@@ -1,0 +1,9 @@
+namespace TurtlePath.Automations.Generation
+{
+    using Microsoft.Extensions.DependencyInjection;
+
+    internal interface IAutomationHandlerTypeGeneratorFactory
+    {
+        IAutomationHandlerTypeGenerator Create(IServiceCollection services);
+    }
+}